+++ src/1_input_parsing/b_validation/validator.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/22 10:54:38 by fschuber          #+#    #+#             */
-/*   Updated: 2024/05/28 23:06:43 by freddy           ###   ########.fr       */
+/*   Updated: 2024/05/28 23:30:30 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,38 +14,57 @@
 
 #define INVALID -1
 
-void flood_fill(t_input_data *data, bool **visited, int x, int y, int *invalid_flag, t_scale map_size)
+int	flood_fill(t_input_data *data, bool **visited, \
+				t_scale pos, t_scale map_size)
 {
-	if (*invalid_flag == INVALID)
-		return;
-	if (x < 0 || y < 0 || x >= map_size.x || y >= map_size.y || *(data->map[y][x]) == VOID)
+	if (pos.x < 0 || pos.y < 0 || pos.x >= map_size.x || \
+		pos.y >= map_size.y || *(data->map[pos.y][pos.x]) == VOID)
+		return (INVALID);
+	if (visited[pos.y][pos.x] == true)
+		return (0);
+	visited[pos.y][pos.x] = true;
+	if (*(data->map[pos.y][pos.x]) == FLOOR)
 	{
-		*invalid_flag = INVALID;
-		return;
+		if (flood_fill(data, visited, (t_scale){pos.x + 1, pos.y}, \
+												map_size) == INVALID)
+			return (INVALID);
+		if (flood_fill(data, visited, (t_scale){pos.x - 1, pos.y}, \
+												map_size) == INVALID)
+			return (INVALID);
+		if (flood_fill(data, visited, (t_scale){pos.x, pos.y + 1}, \
+												map_size) == INVALID)
+			return (INVALID);
+		if (flood_fill(data, visited, (t_scale){pos.x, pos.y - 1}, \
+												map_size) == INVALID)
+			return (INVALID);
 	}
-	if (visited[y][x] == true)
-		return ;
-	visited[y][x] = true;
-	if (*(data->map[y][x]) == FLOOR)
+	return (0);
+}
+
+void	clear_unused_spaces(t_tile_type ***map, \
+						bool **visited, t_scale map_size)
+{
+	int	x;
+	int	y;
+
+	y = -1;
+	while (++y < map_size.y)
 	{
-		flood_fill(data, visited, x + 1, y, invalid_flag, map_size);
-		flood_fill(data, visited, x - 1, y, invalid_flag, map_size);
-		flood_fill(data, visited, x, y + 1, invalid_flag, map_size);
-		flood_fill(data, visited, x, y - 1, invalid_flag, map_size);
+		x = -1;
+		while (++x < map_size.x)
+			if (visited[y][x] == false)
+				*map[y][x] = VOID;
 	}
 }
 
 void	validate(void)
 {
 	t_input_data	*data;
-	int				invalid;
-	int				x;
 	int				y;
 	t_scale			map_size;
 	bool			**visited;
 
 	data = get_persistent_data()->input_data;
-	invalid = 0;
 	map_size.x = 0;
 	map_size.y = 0;
 	while (data->map[map_size.y] != NULL)
@@ -56,16 +75,8 @@ void	validate(void)
 	y = -1;
 	while (++y < map_size.y)
 		visited[y] = gc_malloc(map_size.x * sizeof(bool));
-	printf("y: %f, x: %f\n", get_player()->spawn_point.y, get_player()->spawn_point.x);
-	flood_fill(data, visited, get_player()->spawn_point.x, get_player()->spawn_point.y, &invalid, map_size);
-	if (invalid == INVALID)
-		cub_exit_error("Invalid map. Player can reach edge or void. [Validator]");
-	y = -1;
-	while (++y < map_size.y)
-	{
-		x = -1;
-		while (++x < map_size.x)
-			if (visited[y][x] == false)
-				*data->map[y][x] = VOID;
-	}
+	if (flood_fill(data, visited, (t_scale){get_player()->spawn_point.x, \
+						get_player()->spawn_point.y}, map_size) == INVALID)
+		cub_exit_error("Invalid map - Player can reach edge or void.");
+	clear_unused_spaces(data->map, visited, map_size);
 }